enhancement(build) : rename unit tests files and fix imports
