+++ test/module/middleware.factory.test.ts
@@ -2,7 +2,7 @@ import { normalize } from '@angular-devkit/core';
 import { VirtualTree } from '@angular-devkit/schematics';
 import { SchematicTestRunner, UnitTestTree } from '@angular-devkit/schematics/testing';
 import * as path from 'path';
-import { ApplicationOptions } from '../../src/application/schema';
+import { ApplicationOptions } from '../../src/lib/factories/application.schema';
 import { ModuleOptions } from '../../src/module/schema';
 
 describe('Module Factory', () => {
